% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataverse_manifest.R
\name{access_dataverse_deposit_archive}
\alias{access_dataverse_deposit_archive}
\title{Download and extract a full Dataverse deposit in original format}
\usage{
access_dataverse_deposit_archive(
  dataset,
  deposit_root,
  server = "dataverse.harvard.edu",
  original = TRUE,
  timeout = 3600,
  clean = TRUE
)
}
\description{
Download and extract a full Dataverse deposit in original format
}
\keyword{internal}
