% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code.R
\name{adjust_package_get_code_for_mode}
\alias{adjust_package_get_code_for_mode}
\title{Trim or keep package-repo get_code run section by mode}
\usage{
adjust_package_get_code_for_mode(lines, mode)
}
\description{
Trim or keep package-repo get_code run section by mode
}
\keyword{internal}
