% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code.R
\name{as_replication_code}
\alias{as_replication_code}
\title{Mark script lines so \code{\link[=print.replication_code]{print.replication_code()}} can \code{cat()} them}
\usage{
as_replication_code(lines)
}
\description{
Mark script lines so \code{\link[=print.replication_code]{print.replication_code()}} can \code{cat()} them
}
\keyword{internal}
