% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_replications.R
\name{as_replication_list}
\alias{as_replication_list}
\title{Attach replication_list class and study metadata for printing}
\usage{
as_replication_list(x, doi = NULL, title = NULL)
}
\description{
Attach replication_list class and study metadata for printing
}
\keyword{internal}
