% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code.R
\name{assemble_get_code_run_from_yaml}
\alias{assemble_get_code_run_from_yaml}
\title{Append a yaml-implied run expression (replication.yml is the execute recipe)}
\usage{
assemble_get_code_run_from_yaml(rep, lines)
}
\description{
Append a yaml-implied run expression (replication.yml is the execute recipe)
}
\keyword{internal}
