% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_dag.R
\name{assert_given_outputs_ready}
\alias{assert_given_outputs_ready}
\title{Check that all steps in given set have outputs ready}
\usage{
assert_given_outputs_ready(given_ids, graph, ctx, meta, force = FALSE)
}
\description{
\code{force} is ignored here (same as \code{\link[=assert_parents_ready]{assert_parents_ready()}}): assumed
\code{given} steps must already have outputs on disk.
}
\keyword{internal}
