% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_everything.R
\name{audit_error_snippet}
\alias{audit_error_snippet}
\title{Truncate an error message for audit output}
\usage{
audit_error_snippet(x, max_chars = 240L)
}
\description{
Truncate an error message for audit output
}
\keyword{internal}
