% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_everything.R
\name{audit_job_status_message}
\alias{audit_job_status_message}
\title{Print one audit job progress line with status}
\usage{
audit_job_status_message(label, what, engine, category)
}
\description{
Print one audit job progress line with status
}
\keyword{internal}
