% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_everything.R
\name{audit_progress_categories_from_results}
\alias{audit_progress_categories_from_results}
\title{Vectorized progress categories for an audit results data frame}
\usage{
audit_progress_categories_from_results(results)
}
\description{
Vectorized progress categories for an audit results data frame
}
\keyword{internal}
