% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_everything.R
\name{audit_replication_engine}
\alias{audit_replication_engine}
\title{Infer replication engine from a registry entry}
\usage{
audit_replication_engine(rep)
}
\description{
Infer replication engine from a registry entry
}
\keyword{internal}
