% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_everything.R
\name{audit_replication_group}
\alias{audit_replication_group}
\title{Group id for paired R / Stata replications}
\usage{
audit_replication_group(rep)
}
\description{
Group id for paired R / Stata replications
}
\keyword{internal}
