% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_everything.R
\name{audit_replication_label}
\alias{audit_replication_label}
\title{Display label for a replication entry}
\usage{
audit_replication_label(rep)
}
\description{
Display label for a replication entry
}
\keyword{internal}
