% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/author_display.R
\name{author_name_has_comma_form}
\alias{author_name_has_comma_form}
\title{Whether an author string is already in "Last, First ..." form}
\usage{
author_name_has_comma_form(name)
}
\description{
Whether an author string is already in "Last, First ..." form
}
\keyword{internal}
