% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_requirements.R
\name{check_result}
\alias{check_result}
\title{Record a single checklist result}
\usage{
check_result(name, passed, message = "")
}
\description{
Record a single checklist result
}
\keyword{internal}
