% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_links.R
\name{code_path_language}
\alias{code_path_language}
\title{Infer language for a study-relative code path}
\usage{
code_path_language(rel_path)
}
\description{
Infer language for a study-relative code path
}
\keyword{internal}
