% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_tab_setup.R
\name{code_setup_open_instruction}
\alias{code_setup_open_instruction}
\title{Plain-text instruction for which program to open locally}
\usage{
code_setup_open_instruction(language, study_engines = NULL)
}
\description{
Plain-text instruction for which program to open locally
}
\keyword{internal}
