% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_steps.R
\name{compile_steps_from_legacy}
\alias{compile_steps_from_legacy}
\title{Compile unified steps from legacy prep + replications blocks}
\usage{
compile_steps_from_legacy(meta)
}
\description{
Compile unified steps from legacy prep + replications blocks
}
\keyword{internal}
