% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_replication_template.R
\name{create_replication_template}
\alias{create_replication_template}
\title{Create a replication template}
\usage{
create_replication_template(doi)
}
\arguments{
\item{doi}{Character. DOI of the paper.}
}
\value{
Creates a folder containing replication scaffolding.
}
\description{
Generates a local folder structure for contributing a new replication
to the replication registry. The template includes metadata,
example data, and example scripts.
}
\examples{
\dontrun{
create_replication_template("10.1177/00491241211036161")
}

}
