% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_steps.R
\name{dag_data_roots_node}
\alias{dag_data_roots_node}
\title{Virtual DAG node for raw data roots feeding a step}
\usage{
dag_data_roots_node(paths)
}
\description{
Virtual DAG node for raw data roots feeding a step
}
\keyword{internal}
