% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataverse_manifest.R
\name{dataverse_dataset_archive_url}
\alias{dataverse_dataset_archive_url}
\title{URL for downloading a full dataset archive from Harvard Dataverse}
\usage{
dataverse_dataset_archive_url(
  dataset,
  server = "dataverse.harvard.edu",
  original = TRUE
)
}
\description{
URL for downloading a full dataset archive from Harvard Dataverse
}
\keyword{internal}
