% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declared_data.R
\name{declared_data_entry_url}
\alias{declared_data_entry_url}
\title{Resolve download URL for a declared data entry}
\usage{
declared_data_entry_url(entry, server = "dataverse.harvard.edu")
}
\description{
Resolve download URL for a declared data entry
}
\keyword{internal}
