% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder_replication.R
\docType{data}
\name{DEFAULT_REGISTRY_REPO}
\alias{DEFAULT_REGISTRY_REPO}
\title{Default registry repository slug}
\format{
An object of class \code{character} of length 1.
}
\usage{
DEFAULT_REGISTRY_REPO
}
\description{
Default registry repository slug
}
\keyword{internal}
