% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code.R
\name{detect_analysis_fn_name}
\alias{detect_analysis_fn_name}
\title{Analysis function name defined in script lines (make_* or generate_*)}
\usage{
detect_analysis_fn_name(lines, what)
}
\description{
Analysis function name defined in script lines (make_* or generate_*)
}
\keyword{internal}
