% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code.R
\name{detect_format_fn_name}
\alias{detect_format_fn_name}
\title{Format helper name from yaml or script definitions}
\usage{
detect_format_fn_name(rep, lines = NULL)
}
\description{
Format helper name from yaml or script definitions
}
\keyword{internal}
