% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_requirements.R
\docType{data}
\name{DISPLAY_REPLICATION_TYPES}
\alias{DISPLAY_REPLICATION_TYPES}
\title{Display replication types validated for artifacts}
\format{
An object of class \code{character} of length 2.
}
\usage{
DISPLAY_REPLICATION_TYPES
}
\description{
Display replication types validated for artifacts
}
\keyword{internal}
