% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_requirements.R
\name{doi_to_registry_folder}
\alias{doi_to_registry_folder}
\title{Registry folder name from a DOI}
\usage{
doi_to_registry_folder(doi)
}
\description{
Registry folder name from a DOI
}
\keyword{internal}
