% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_tab_setup.R
\name{engine_display_name}
\alias{engine_display_name}
\title{Display name for a replication engine}
\usage{
engine_display_name(engine)
}
\description{
Display name for a replication engine
}
\keyword{internal}
