% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_replication.R
\name{ensure_replication_package}
\alias{ensure_replication_package}
\title{Ensure a study replication package is installed}
\usage{
ensure_replication_package(package, repo)
}
\arguments{
\item{package}{R package name.}

\item{repo}{GitHub slug \code{org/repo}.}
}
\description{
Ensure a study replication package is installed
}
\keyword{internal}
