% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata_replication.R
\name{ensure_stata_available}
\alias{ensure_stata_available}
\title{Ensure Stata is available for a replication entry}
\usage{
ensure_stata_available(rep)
}
\arguments{
\item{rep}{Replication entry.}
}
\description{
Ensure Stata is available for a replication entry
}
\keyword{internal}
