% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_hints.R
\name{executable_renviron_lines}
\alias{executable_renviron_lines}
\title{Suggest permanent Python / Stata paths via Renviron}
\usage{
executable_renviron_lines()
}
\value{
Character vector of \code{KEY=value} lines.
}
\description{
Suggest permanent Python / Stata paths via Renviron
}
\keyword{internal}
