% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_inheritance.R
\name{extended_study_base_root}
\alias{extended_study_base_root}
\title{Local root of the base study for an extension, if any}
\usage{
extended_study_base_root(meta)
}
\description{
Local root of the base study for an extension, if any
}
\keyword{internal}
