% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_links.R
\name{extract_r_source_calls}
\alias{extract_r_source_calls}
\title{Extract R \code{source()} calls from code lines}
\usage{
extract_r_source_calls(lines)
}
\description{
Extract R \code{source()} calls from code lines
}
\keyword{internal}
