% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_replications.R
\name{filter_replication_entries}
\alias{filter_replication_entries}
\title{Filter step entries by display / pipeline / all}
\usage{
filter_replication_entries(meta, include = c("display", "pipeline", "all"))
}
\description{
Filter step entries by display / pipeline / all
}
\keyword{internal}
