% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_replication.R
\name{find_package_yaml_entry}
\alias{find_package_yaml_entry}
\title{Find a prep/replication entry by id in package yaml}
\usage{
find_package_yaml_entry(meta, id)
}
\description{
Find a prep/replication entry by id in package yaml
}
\keyword{internal}
