% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_replication.R
\name{find_prep_entry}
\alias{find_prep_entry}
\title{Find a prep entry by id}
\usage{
find_prep_entry(meta, step_id)
}
\arguments{
\item{meta}{Parsed replication metadata.}

\item{step_id}{Prep step identifier.}
}
\description{
Find a prep entry by id
}
\keyword{internal}
