% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_replication.R
\name{find_python_executable}
\alias{find_python_executable}
\title{Find a Python executable}
\usage{
find_python_executable()
}
\value{
Character path.
}
\description{
Find a Python executable
}
\keyword{internal}
