% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder_requirements.R
\name{folder_display_replications}
\alias{folder_display_replications}
\title{Display replications from study yaml}
\usage{
folder_display_replications(meta)
}
\description{
Display replications from study yaml
}
\keyword{internal}
