% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder_requirements.R
\name{folder_study_run_options}
\alias{folder_study_run_options}
\title{Options for running replications against a local folder-backed study}
\usage{
folder_study_run_options(study_root, meta, registry_root = NULL)
}
\description{
Options for running replications against a local folder-backed study
}
\keyword{internal}
