% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/author_display.R
\name{format_author_name}
\alias{format_author_name}
\title{Format one author as \code{Last, First [Middle ...]}}
\usage{
format_author_name(name)
}
\description{
Format one author as \code{Last, First [Middle ...]}
}
\keyword{internal}
