% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/author_display.R
\name{format_authors_summary}
\alias{format_authors_summary}
\title{Full author list for study details panel}
\usage{
format_authors_summary(authors_str)
}
\description{
Full author list for study details panel
}
\keyword{internal}
