% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_dag.R
\name{format_child_for_step}
\alias{format_child_for_step}
\title{Format child id for a step, if registered}
\usage{
format_child_for_step(step_id, graph)
}
\description{
Format child id for a step, if registered
}
\keyword{internal}
