% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_steps.R
\name{format_child_step}
\alias{format_child_step}
\title{Format child step for an analytical step id, if any}
\usage{
format_child_step(meta, step_id)
}
\description{
Format child step for an analytical step id, if any
}
\keyword{internal}
