% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_links.R
\name{format_code_link_diagnostic_title}
\alias{format_code_link_diagnostic_title}
\title{Human-readable tooltip for a failed code link}
\usage{
format_code_link_diagnostic_title(diag)
}
\description{
Human-readable tooltip for a failed code link
}
\keyword{internal}
