% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_replication.R
\name{get_code_from_package_repo}
\alias{get_code_from_package_repo}
\title{Build get_code output for a package-backed study from remote source}
\usage{
get_code_from_package_repo(
  meta,
  ctx,
  what,
  package,
  mode = c("definitions", "run")
)
}
\arguments{
\item{meta}{Parsed replication metadata (with replications list).}

\item{ctx}{Paper context.}

\item{what}{Replication id.}

\item{package}{R package name.}
}
\description{
Build get_code output for a package-backed study from remote source
}
\keyword{internal}
