% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code.R
\name{get_code_result_kind}
\alias{get_code_result_kind}
\title{Noun for get_code / Code-tab run tips (table / figure / step / result)}
\usage{
get_code_result_kind(type = NULL)
}
\description{
Noun for get_code / Code-tab run tips (table / figure / step / result)
}
\keyword{internal}
