% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code.R
\name{get_code_tip_call_args}
\alias{get_code_tip_call_args}
\title{Quoted args for run_replication / get_code examples}
\usage{
get_code_tip_call_args(doi = NULL, what = NULL)
}
\description{
Quoted args for run_replication / get_code examples
}
\keyword{internal}
