% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_doi_metadata.R
\name{get_doi_metadata}
\alias{get_doi_metadata}
\title{Get DOI metadata}
\usage{
get_doi_metadata(doi)
}
\arguments{
\item{doi}{Character DOI of the paper.}
}
\description{
Retrieves metadata for a paper from its DOI.
}
