% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_replication.R
\name{get_study_package_object}
\alias{get_study_package_object}
\title{Load a LazyData / exported object from a study package}
\usage{
get_study_package_object(name, package)
}
\description{
Load a LazyData / exported object from a study package
}
\keyword{internal}
