% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_replication.R
\name{github_package_outdated}
\alias{github_package_outdated}
\title{Whether an installed package lags the GitHub ref}
\usage{
github_package_outdated(package, repo, ref = "main")
}
\arguments{
\item{package}{Installed package name.}

\item{repo}{GitHub slug \code{org/repo}.}

\item{ref}{Branch, tag, or commit.}
}
\description{
Whether an installed package lags the GitHub ref
}
\keyword{internal}
