% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code.R
\name{has_nframe_footer}
\alias{has_nframe_footer}
\title{Whether script lines contain an interactive sys.nframe() footer (legacy)}
\usage{
has_nframe_footer(lines)
}
\description{
Whether script lines contain an interactive sys.nframe() footer (legacy)
}
\keyword{internal}
