% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substantive_checks.R
\name{has_substantive_check}
\alias{has_substantive_check}
\title{Whether a study defines a substantive check for a replication step}
\usage{
has_substantive_check(study_root, what)
}
\description{
Whether a study defines a substantive check for a replication step
}
\keyword{internal}
