% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_hints.R
\name{hint_quote_study}
\alias{hint_quote_study}
\title{Format a DOI or path for use in maintainer hint code}
\usage{
hint_quote_study(doi)
}
\description{
Format a DOI or path for use in maintainer hint code
}
\keyword{internal}
