% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_replication.R
\name{html_escape_text}
\alias{html_escape_text}
\title{Escape text for safe inclusion in HTML}
\usage{
html_escape_text(text)
}
\description{
Escape text for safe inclusion in HTML
}
\keyword{internal}
