% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code.R
\name{infer_stata_source_paths}
\alias{infer_stata_source_paths}
\title{Infer substantive Stata scripts from nested \code{do} calls in a runner.}
\usage{
infer_stata_source_paths(wrapper_lines)
}
\description{
Infer substantive Stata scripts from nested \code{do} calls in a runner.
}
\keyword{internal}
