% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_replication.R
\name{install_replication_package_github}
\alias{install_replication_package_github}
\title{Install or upgrade a study package from GitHub}
\usage{
install_replication_package_github(package, repo, ref = "main")
}
\arguments{
\item{package}{R package name.}

\item{repo}{GitHub slug.}

\item{ref}{Git ref.}
}
\description{
Install or upgrade a study package from GitHub
}
\keyword{internal}
