% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maintainer_dependencies.R
\name{install_study_stata_packages}
\alias{install_study_stata_packages}
\title{Install Stata SSC / GitHub packages declared for a study}
\usage{
install_study_stata_packages(study_root, meta)
}
\description{
Install Stata SSC / GitHub packages declared for a study
}
\keyword{internal}
