% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/author_display.R
\name{is_author_middle_initial}
\alias{is_author_middle_initial}
\title{Detect Stata-style middle initials in author names}
\usage{
is_author_middle_initial(part)
}
\description{
Detect Stata-style middle initials in author names
}
\keyword{internal}
