% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_steps.R
\name{is_display_step_type}
\alias{is_display_step_type}
\title{Whether a step type is a display sink (table / figure)}
\usage{
is_display_step_type(type)
}
\description{
Whether a step type is a display sink (table / figure)
}
\keyword{internal}
