% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_inheritance.R
\name{is_inherited_step}
\alias{is_inherited_step}
\title{Whether a step runs in the base study repository}
\usage{
is_inherited_step(step)
}
\description{
Whether a step runs in the base study repository
}
\keyword{internal}
