% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_replication.R
\name{is_package_replication}
\alias{is_package_replication}
\title{Whether replication metadata refers to an installed R package}
\usage{
is_package_replication(meta)
}
\arguments{
\item{meta}{Parsed replication.yml contents.}
}
\value{
Logical.
}
\description{
Whether replication metadata refers to an installed R package
}
