% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_steps.R
\name{is_pipeline_step_type}
\alias{is_pipeline_step_type}
\title{Whether a step type is a pipeline / transform step}
\usage{
is_pipeline_step_type(type)
}
\description{
Whether a step type is a pipeline / transform step
}
\keyword{internal}
