% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_replication.R
\name{is_prep_entry}
\alias{is_prep_entry}
\title{Whether a yaml entry is a pipeline / prep step}
\usage{
is_prep_entry(rep)
}
\arguments{
\item{rep}{Replication or prep entry from \code{replication.yml}.}
}
\value{
Logical scalar.
}
\description{
Whether a yaml entry is a pipeline / prep step
}
\keyword{internal}
