% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_replication.R
\name{is_study_package_error}
\alias{is_study_package_error}
\title{Whether an error is a missing study replication package}
\usage{
is_study_package_error(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Logical.
}
\description{
Whether an error is a missing study replication package
}
\keyword{internal}
