% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder_replication.R
\name{is_study_path_query}
\alias{is_study_path_query}
\title{Detect whether input is a filesystem path to a study repository}
\usage{
is_study_path_query(x)
}
\arguments{
\item{x}{Character scalar.}
}
\value{
Logical scalar.
}
\description{
Detect whether input is a filesystem path to a study repository
}
\keyword{internal}
