% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_links.R
\name{is_study_root_usable}
\alias{is_study_root_usable}
\title{Whether a study root path is usable for on-disk code linking}
\usage{
is_study_root_usable(study_root)
}
\description{
Whether a study root path is usable for on-disk code linking
}
\keyword{internal}
