% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_tabular_to_html.R
\name{latex_inline_to_text}
\alias{latex_inline_to_text}
\title{Plain-text version of inline LaTeX (for captions).}
\usage{
latex_inline_to_text(text)
}
\description{
Plain-text version of inline LaTeX (for captions).
}
\keyword{internal}
