% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_tabular_to_html.R
\name{latex_split_rows}
\alias{latex_split_rows}
\title{Split tabular body on row-ending \\\\ outside braces.}
\usage{
latex_split_rows(body)
}
\description{
Split tabular body on row-ending \\\\ outside braces.
}
\keyword{internal}
