% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_replications.R
\name{list_replication_groups_impl}
\alias{list_replication_groups_impl}
\title{One replication entry per logical group (internal)}
\usage{
list_replication_groups_impl(
  meta,
  entries = NULL,
  language = NULL,
  display_only = TRUE
)
}
\description{
One replication entry per logical group (internal)
}
\keyword{internal}
