% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_inheritance.R
\name{load_extended_base_meta}
\alias{load_extended_base_meta}
\title{Load base study replication.yml for an extension}
\usage{
load_extended_base_meta(extends)
}
\description{
Load base study replication.yml for an extension
}
\keyword{internal}
