% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_replication.R
\name{load_replication_package_path}
\alias{load_replication_package_path}
\title{Load a package from a local source tree}
\usage{
load_replication_package_path(path, package)
}
\arguments{
\item{path}{Path to package root (contains DESCRIPTION).}

\item{package}{Expected package name.}
}
\description{
Load a package from a local source tree
}
\keyword{internal}
