% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_hints.R
\name{maintainer_dependency_hint_ui}
\alias{maintainer_dependency_hint_ui}
\title{Shiny-friendly list form of \code{\link[=maintainer_dependency_hint]{maintainer_dependency_hint()}}}
\usage{
maintainer_dependency_hint_ui(...)
}
\arguments{
\item{...}{Passed to \code{\link[=maintainer_dependency_hint]{maintainer_dependency_hint()}}.}
}
\value{
List with \code{title}, \code{body}, and \code{lines}.
}
\description{
Shiny-friendly list form of \code{\link[=maintainer_dependency_hint]{maintainer_dependency_hint()}}
}
\keyword{internal}
