% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_replication_functions.R
\name{make_function_name}
\alias{make_function_name}
\title{Standard make_* function name for a replication id}
\usage{
make_function_name(what)
}
\arguments{
\item{what}{Replication identifier.}
}
\description{
Standard make_* function name for a replication id
}
\keyword{internal}
