% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataverse_manifest.R
\name{manifest_row_use_original}
\alias{manifest_row_use_original}
\title{Whether a manifest row should use Dataverse original-format download}
\usage{
manifest_row_use_original(row)
}
\description{
Whether a manifest row should use Dataverse original-format download
}
\keyword{internal}
