% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_inheritance.R
\name{merge_extended_study_meta}
\alias{merge_extended_study_meta}
\title{Merge extension metadata with its base study}
\usage{
merge_extended_study_meta(meta, ctx = NULL)
}
\description{
Merge extension metadata with its base study
}
\keyword{internal}
