% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_inheritance.R
\name{merge_extended_study_steps}
\alias{merge_extended_study_steps}
\title{Expand inherit entries and merge with base steps}
\usage{
merge_extended_study_steps(extension_meta, base_meta, extends)
}
\description{
Expand inherit entries and merge with base steps
}
\keyword{internal}
