% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_replication.R
\name{missing_data_message}
\alias{missing_data_message}
\title{Message when an output cannot be shown or re-run due to unavailable data}
\usage{
missing_data_message(
  output,
  data_class,
  mode = c("not_available", "not_reproducible")
)
}
\description{
Message when an output cannot be shown or re-run due to unavailable data
}
\keyword{internal}
