% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder_replication.R
\name{move_directory}
\alias{move_directory}
\title{Move or copy a directory to a destination path}
\usage{
move_directory(from, to)
}
\arguments{
\item{from}{Source directory.}

\item{to}{Destination directory.}
}
\description{
Move or copy a directory to a destination path
}
\keyword{internal}
