% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_html_table.R
\name{normalize_html_table}
\alias{normalize_html_table}
\title{Decode HTML entities in kableExtra / modelsummary table output}
\usage{
normalize_html_table(html)
}
\arguments{
\item{html}{Character HTML string.}
}
\value{
Character HTML with common entities decoded for browser display.
}
\description{
Decode HTML entities in kableExtra / modelsummary table output
}
\examples{
\dontrun{
normalize_html_table("&lt;table&gt;&amp;nbsp;data&lt;/table&gt;")
}

}
\keyword{internal}
