% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_links.R
\name{normalize_path_slashes}
\alias{normalize_path_slashes}
\title{Normalize path separators to forward slashes}
\usage{
normalize_path_slashes(path)
}
\description{
Normalize path separators to forward slashes
}
\keyword{internal}
