% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_steps.R
\name{output_path_ready}
\alias{output_path_ready}
\title{Whether a path exists and is non-empty (directory must contain files)}
\usage{
output_path_ready(path)
}
\description{
Whether a path exists and is non-empty (directory must contain files)
}
\keyword{internal}
