% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_build_info.R
\name{package_library_path}
\alias{package_library_path}
\title{Installed package library path}
\usage{
package_library_path(package = "replicateEverything")
}
\arguments{
\item{package}{Package name.}
}
\value{
Normalized path to the installed package directory, or \code{""}.
}
\description{
Installed package library path
}
\keyword{internal}
