% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_registry.R
\name{package_name_from_root}
\alias{package_name_from_root}
\title{Package name from a local package source tree}
\usage{
package_name_from_root(pkg_root)
}
\description{
Package name from a local package source tree
}
\keyword{internal}
