% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_build_info.R
\name{package_namespace_stale}
\alias{package_namespace_stale}
\title{Whether a loaded package namespace is older than the on-disk install}
\usage{
package_namespace_stale(package = "replicateEverything")
}
\arguments{
\item{package}{Package name.}
}
\value{
Logical scalar.
}
\description{
Whether a loaded package namespace is older than the on-disk install
}
\keyword{internal}
