% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_build_info.R
\name{package_remote_install_sha}
\alias{package_remote_install_sha}
\title{GitHub install SHA from package DESCRIPTION}
\usage{
package_remote_install_sha(package = "replicateEverything")
}
\arguments{
\item{package}{Package name.}
}
\value{
Seven-character SHA or \code{NA_character_}.
}
\description{
GitHub install SHA from package DESCRIPTION
}
\keyword{internal}
