% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_requirements.R
\docType{data}
\name{PACKAGE_REPLICATION_API}
\alias{PACKAGE_REPLICATION_API}
\title{Required exported API for package-backed replication studies}
\format{
An object of class \code{character} of length 7.
}
\usage{
PACKAGE_REPLICATION_API
}
\description{
Required exported API for package-backed replication studies
}
\keyword{internal}
