% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_replication.R
\name{package_repo_ref}
\alias{package_repo_ref}
\title{Git ref for \code{remotes::install_github()}}
\usage{
package_repo_ref(meta)
}
\arguments{
\item{meta}{Parsed replication.yml contents.}
}
\value{
Character branch, tag, or commit.
}
\description{
Git ref for \code{remotes::install_github()}
}
\keyword{internal}
