% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_artifact_helpers.R
\name{package_source_root}
\alias{package_source_root}
\title{Installed package source tree}
\usage{
package_source_root(package)
}
\description{
Installed package source tree
}
\keyword{internal}
