% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_replication.R
\name{package_yaml_entries}
\alias{package_yaml_entries}
\title{Package yaml entries (prep + replications + steps)}
\usage{
package_yaml_entries(meta)
}
\description{
Package yaml entries (prep + replications + steps)
}
\keyword{internal}
