% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_inheritance.R
\name{parse_inherit_reference}
\alias{parse_inherit_reference}
\title{Parse an inherit reference (\code{repo/step} or \code{step})}
\usage{
parse_inherit_reference(ref, default_repo = NULL)
}
\description{
Parse an inherit reference (\code{repo/step} or \code{step})
}
\keyword{internal}
