% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_links.R
\name{parse_stata_globals}
\alias{parse_stata_globals}
\title{Parse \code{global} assignments from Stata lines}
\usage{
parse_stata_globals(lines, existing = character())
}
\description{
Parse \code{global} assignments from Stata lines
}
\keyword{internal}
