% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_replication.R
\name{prep_manifest_output_path}
\alias{prep_manifest_output_path}
\title{Relative path for prep output in build manifests}
\usage{
prep_manifest_output_path(path, root = NULL)
}
\description{
Relative path for prep output in build manifests
}
\keyword{internal}
