% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_links.R
\name{prepare_code_viewer_state}
\alias{prepare_code_viewer_state}
\title{Prepare code viewer state for Shiny}
\usage{
prepare_code_viewer_state(
  doi,
  what,
  language = NULL,
  repo = NULL,
  folder = NULL
)
}
\value{
List suitable for serializing to the client.
}
\description{
Prepare code viewer state for Shiny
}
\keyword{internal}
