% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_study_plan.R
\name{prepare_study_run}
\alias{prepare_study_run}
\title{Prepare and validate a study run plan}
\usage{
prepare_study_run(
  doi,
  what,
  given = "parents",
  format = FALSE,
  force = FALSE,
  repo = NULL,
  folder = NULL
)
}
\description{
Prepare and validate a study run plan
}
\keyword{internal}
