% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_replication.R
\name{preview_data_file}
\alias{preview_data_file}
\title{Preview a processed data file (head rows)}
\usage{
preview_data_file(path, n = 6L)
}
\arguments{
\item{path}{Local file path.}

\item{n}{Maximum rows to read.}
}
\description{
Preview a processed data file (head rows)
}
\keyword{internal}
