% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code.R
\name{print.replication_code}
\alias{print.replication_code}
\title{Print replication script lines as code, not a quoted vector}
\usage{
\method{print}{replication_code}(x, ...)
}
\arguments{
\item{x}{A \code{replication_code} object from \code{\link[=get_code]{get_code()}}.}

\item{...}{Ignored.}
}
\description{
Print replication script lines as code, not a quoted vector
}
\keyword{internal}
