% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_replications.R
\name{print.replication_list}
\alias{print.replication_list}
\title{Compact print method for replication step lists}
\usage{
\method{print}{replication_list}(x, n = 20, ...)
}
\arguments{
\item{x}{A \code{replication_list} object.}

\item{n}{Maximum rows to display.}

\item{...}{Ignored.}
}
\description{
Compact print method for replication step lists
}
