% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_replication.R
\name{python_executable_candidates}
\alias{python_executable_candidates}
\title{Candidate Python executables}
\usage{
python_executable_candidates()
}
\value{
Character vector of paths.
}
\description{
Candidate Python executables
}
\keyword{internal}
