% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_replication.R
\name{python_run_dir}
\alias{python_run_dir}
\title{Working directory for a Python replication run}
\usage{
python_run_dir(rep, ctx, meta = NULL)
}
\description{
Working directory for a Python replication run
}
\keyword{internal}
