% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_replication.R
\name{raw_to_github_browse}
\alias{raw_to_github_browse}
\title{Convert a raw GitHub URL to a browseable \code{/blob/} link}
\usage{
raw_to_github_browse(url)
}
\arguments{
\item{url}{Raw \code{raw.githubusercontent.com} URL.}
}
\value{
Browse URL or original \code{url}.
}
\description{
Convert a raw GitHub URL to a browseable \code{/blob/} link
}
\keyword{internal}
