% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_build_info.R
\name{read_build_sha_file}
\alias{read_build_sha_file}
\title{Read first line from a build stamp file}
\usage{
read_build_sha_file(path)
}
\arguments{
\item{path}{File path.}
}
\description{
Read first line from a build stamp file
}
\keyword{internal}
