% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_replication.R
\name{read_installed_package_source}
\alias{read_installed_package_source}
\title{Read make_*/prep source from an installed study package}
\usage{
read_installed_package_source(name, package)
}
\description{
Read make_*/prep source from an installed study package
}
\keyword{internal}
