% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_replication.R
\name{read_lines_url}
\alias{read_lines_url}
\title{Read text lines from a local path or HTTP(S) URL}
\usage{
read_lines_url(url)
}
\arguments{
\item{url}{Character URL or file path.}
}
\value{
Character vector of lines, or empty if not found.
}
\description{
Read text lines from a local path or HTTP(S) URL
}
\keyword{internal}
