% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_artifact_helpers.R
\name{read_package_replication_meta}
\alias{read_package_replication_meta}
\title{Read replication metadata from an installed study package}
\usage{
read_package_replication_meta(package)
}
\description{
Read replication metadata from an installed study package
}
\keyword{internal}
