% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_requirements.R
\name{read_package_replication_yaml}
\alias{read_package_replication_yaml}
\title{Read replication yaml from a package source tree}
\usage{
read_package_replication_yaml(pkg_root)
}
\description{
Read replication yaml from a package source tree
}
\keyword{internal}
