% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_registry.R
\name{read_study_meta_from_root}
\alias{read_study_meta_from_root}
\title{Read full replication metadata from a study repository root}
\usage{
read_study_meta_from_root(study_root, kind = NULL)
}
\description{
Read full replication metadata from a study repository root
}
\keyword{internal}
