% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder_requirements.R
\name{read_study_replication_yaml}
\alias{read_study_replication_yaml}
\title{Read replication yaml from a folder-backed study tree}
\usage{
read_study_replication_yaml(study_root)
}
\description{
Read replication yaml from a folder-backed study tree
}
\keyword{internal}
