% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_registry.R
\name{registry_folder_from_paper}
\alias{registry_folder_from_paper}
\title{Registry folder / handle from paper metadata}
\usage{
registry_folder_from_paper(paper)
}
\description{
Registry folder / handle from paper metadata
}
\keyword{internal}
