% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder_replication.R
\name{registry_paper_yaml_url}
\alias{registry_paper_yaml_url}
\title{GitHub raw URL for a registry paper stub yaml}
\usage{
registry_paper_yaml_url(
  folder,
  registry_repo = DEFAULT_REGISTRY_REPO,
  ref = "main"
)
}
\arguments{
\item{folder}{Registry folder name.}

\item{registry_repo}{Registry repository slug.}

\item{ref}{Git ref.}
}
\description{
GitHub raw URL for a registry paper stub yaml
}
\keyword{internal}
