% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder_replication.R
\name{registry_papers_subdir_legacy}
\alias{registry_papers_subdir_legacy}
\title{Legacy registry subdirectory (pre-rename)}
\usage{
registry_papers_subdir_legacy()
}
\description{
Legacy registry subdirectory (pre-rename)
}
\keyword{internal}
