% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_requirements.R
\name{registry_stub_from_package_meta}
\alias{registry_stub_from_package_meta}
\title{Build the lightweight registry stub yaml list from package metadata}
\usage{
registry_stub_from_package_meta(meta, package_folder = NULL)
}
\description{
Build the lightweight registry stub yaml list from package metadata
}
\keyword{internal}
