% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder_replication.R
\name{registry_studies_dir}
\alias{registry_studies_dir}
\title{Absolute path to \code{registry/studies/}}
\usage{
registry_studies_dir(registry_root)
}
\arguments{
\item{registry_root}{Registry checkout root.}
}
\description{
Absolute path to \code{registry/studies/}
}
\keyword{internal}
