% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder_replication.R
\name{registry_studies_subdir}
\alias{registry_studies_subdir}
\title{Registry subdirectory name for study stub yaml files}
\usage{
registry_studies_subdir()
}
\description{
Registry subdirectory name for study stub yaml files
}
\keyword{internal}
