% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder_replication.R
\name{registry_study_yaml_path}
\alias{registry_study_yaml_path}
\title{Path to a registry study stub yaml file}
\usage{
registry_study_yaml_path(registry_root, folder)
}
\arguments{
\item{registry_root}{Registry checkout root.}

\item{folder}{Registry folder name.}
}
\value{
Character path under \code{studies/<folder>.yml}.
}
\description{
Path to a registry study stub yaml file
}
\keyword{internal}
