% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder_replication.R
\name{registry_study_yaml_url}
\alias{registry_study_yaml_url}
\title{GitHub raw URL for a registry study stub yaml}
\usage{
registry_study_yaml_url(
  folder,
  registry_repo = DEFAULT_REGISTRY_REPO,
  ref = "main"
)
}
\arguments{
\item{folder}{Registry folder name.}

\item{registry_repo}{Registry repository slug.}

\item{ref}{Git ref.}
}
\description{
GitHub raw URL for a registry study stub yaml
}
\keyword{internal}
