% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_replication.R
\name{registry_url}
\alias{registry_url}
\title{Join a registry base URL with a relative path without duplicate slashes.}
\usage{
registry_url(base, rel)
}
\arguments{
\item{base}{Character base URL.}

\item{rel}{Character relative path.}
}
\description{
Join a registry base URL with a relative path without duplicate slashes.
}
\keyword{internal}
