% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_related.R
\name{related_studies_for_index_row}
\alias{related_studies_for_index_row}
\title{Related studies for one index row}
\usage{
related_studies_for_index_row(row, index = NULL)
}
\description{
Related studies for one index row
}
\keyword{internal}
