% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_replication.R
\name{replication_code_language_python}
\alias{replication_code_language_python}
\title{Code language for Python entries}
\usage{
replication_code_language_python()
}
\description{
Code language for Python entries
}
\keyword{internal}
