% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replication_resolve.R
\name{replication_logical_id}
\alias{replication_logical_id}
\title{Logical replication id (group) for a yaml entry}
\usage{
replication_logical_id(rep)
}
\arguments{
\item{rep}{Replication entry from \code{replication.yml}.}
}
\value{
Character scalar.
}
\description{
Logical replication id (group) for a yaml entry
}
\keyword{internal}
