% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_dag.R
\name{resolve_given_set}
\alias{resolve_given_set}
\title{Expand given to a set of step ids assumed complete}
\usage{
resolve_given_set(given, target_id, graph)
}
\description{
Expand given to a set of step ids assumed complete
}
\keyword{internal}
