% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_requirements.R
\name{resolve_package_location}
\alias{resolve_package_location}
\title{Resolve a study package root directory from a local path or GitHub address}
\usage{
resolve_package_location(location)
}
\arguments{
\item{location}{Local package path or GitHub address.}
}
\value{
Normalized path to package root.
}
\description{
Resolve a study package root directory from a local path or GitHub address
}
\keyword{internal}
