% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_links.R
\name{resolve_stata_path}
\alias{resolve_stata_path}
\title{Resolve a Stata path (alias for \code{resolve_code_path})}
\usage{
resolve_stata_path(path, study_root, globals = character(), from_file = NULL)
}
\description{
Resolve a Stata path (alias for \code{resolve_code_path})
}
\keyword{internal}
