% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_replication.R
\name{run_python_replication}
\alias{run_python_replication}
\title{Execute a Python script or notebook}
\usage{
run_python_replication(rep, ctx, meta = NULL)
}
\arguments{
\item{rep}{Replication entry.}

\item{ctx}{Paper context.}

\item{meta}{Parsed metadata.}
}
\description{
Execute a Python script or notebook
}
\keyword{internal}
