% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata_replication.R
\name{run_stata_do}
\alias{run_stata_do}
\title{Run a Stata do-file non-interactively}
\usage{
run_stata_do(do_path, workdir, timeout = 900L)
}
\arguments{
\item{do_path}{Path to the do-file.}

\item{workdir}{Working directory Stata should use.}

\item{timeout}{Seconds before aborting (best effort on Windows).}
}
\description{
Run a Stata do-file non-interactively
}
\keyword{internal}
