% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_papers.R
\name{search_papers}
\alias{search_papers}
\title{Search replicated papers}
\usage{
search_papers(query)
}
\arguments{
\item{query}{Character string used to search paper titles, handles, or authors.}
}
\value{
A filtered data frame of matching papers.
}
\description{
Searches the replication registry for papers matching a keyword
in the title, handle, or author fields.
}
\examples{
\dontrun{
search_papers("insurgency")
search_papers("Blair")
search_papers("Acemoglu")
search_papers("Fearon")
search_papers("S0003055403000534")
search_papers("rep-template")
}

}
