% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_feedback.R
\name{SHINY_FEEDBACK_CATEGORIES}
\alias{SHINY_FEEDBACK_CATEGORIES}
\title{Allowed Shiny feedback categories}
\usage{
SHINY_FEEDBACK_CATEGORIES
}
\description{
Allowed Shiny feedback categories
}
\keyword{internal}
