% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_feedback.R
\name{shiny_feedback_category_meta}
\alias{shiny_feedback_category_meta}
\title{Label and title prefix for a feedback category}
\usage{
shiny_feedback_category_meta(category)
}
\arguments{
\item{category}{Allowlisted category.}
}
\value{
Named list with \code{label} and \code{title_prefix}.
}
\description{
Label and title prefix for a feedback category
}
\keyword{internal}
