% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_feedback.R
\name{SHINY_FEEDBACK_GITHUB_REPO}
\alias{SHINY_FEEDBACK_GITHUB_REPO}
\title{Default GitHub repo for Shiny feedback issue links}
\usage{
SHINY_FEEDBACK_GITHUB_REPO
}
\description{
Default GitHub repo for Shiny feedback issue links
}
\keyword{internal}
