% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_feedback.R
\name{shiny_feedback_log_enabled}
\alias{shiny_feedback_log_enabled}
\title{Whether server-side Shiny feedback logging is enabled}
\usage{
shiny_feedback_log_enabled()
}
\value{
Logical scalar.
}
\description{
Whether server-side Shiny feedback logging is enabled
}
\keyword{internal}
