% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_build_info.R
\name{short_build_sha}
\alias{short_build_sha}
\title{Short git or GitHub SHA label}
\usage{
short_build_sha(sha)
}
\arguments{
\item{sha}{Full or short commit SHA.}
}
\value{
Seven-character SHA or \code{NA_character_}.
}
\description{
Short git or GitHub SHA label
}
\keyword{internal}
