% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_replication.R
\name{sibling_monorepo_root}
\alias{sibling_monorepo_root}
\title{Parent directory of the registry when developing in a monorepo}
\usage{
sibling_monorepo_root()
}
\value{
Character path or \code{NULL}.
}
\description{
Parent directory of the registry when developing in a monorepo
}
\keyword{internal}
