% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata_replication.R
\name{stata_deps_install_lines_from_packages}
\alias{stata_deps_install_lines_from_packages}
\title{Build a Stata install script from \code{stata_packages:}}
\usage{
stata_deps_install_lines_from_packages(packages)
}
\arguments{
\item{packages}{Character vector of SSC ado names.}
}
\value{
Character vector of Stata commands.
}
\description{
Build a Stata install script from \code{stata_packages:}
}
\keyword{internal}
