% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata_replication.R
\name{stata_deps_probe_code_map}
\alias{stata_deps_probe_code_map}
\title{Exit-code -> package name map for \code{\link[=stata_deps_probe_lines_from_packages]{stata_deps_probe_lines_from_packages()}}}
\usage{
stata_deps_probe_code_map(packages)
}
\arguments{
\item{packages}{Character vector of ado command names.}
}
\value{
Named character vector (exit code as name, package as value).
}
\description{
Exit-code -> package name map for \code{\link[=stata_deps_probe_lines_from_packages]{stata_deps_probe_lines_from_packages()}}
}
\keyword{internal}
