% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata_replication.R
\name{stata_package_probe_line}
\alias{stata_package_probe_line}
\title{Stata do-file line(s) that test presence of a declared package}
\usage{
stata_package_probe_line(pkg)
}
\arguments{
\item{pkg}{Declared package name from \code{stata_packages:}.}
}
\value{
Character scalar: a single \code{cap which ...} or
\code{cap findfile ...} line.
}
\description{
Stata do-file line(s) that test presence of a declared package
}
\keyword{internal}
