% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_dag.R
\name{step_ancestors}
\alias{step_ancestors}
\title{Ancestors of a step (transitive parents, excluding format nodes as parents)}
\usage{
step_ancestors(step_id, graph)
}
\description{
Ancestors of a step (transitive parents, excluding format nodes as parents)
}
\keyword{internal}
