% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_steps.R
\name{step_declared_input_paths}
\alias{step_declared_input_paths}
\title{Relative input paths declared for a step}
\usage{
step_declared_input_paths(step)
}
\description{
Relative input paths declared for a step
}
\keyword{internal}
