% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_dag.R
\name{step_direct_parents}
\alias{step_direct_parents}
\title{Direct analytical parents (excludes format)}
\usage{
step_direct_parents(step_id, graph)
}
\description{
Direct analytical parents (excludes format)
}
\keyword{internal}
