% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_links.R
\name{strip_stata_comments}
\alias{strip_stata_comments}
\title{Strip Stata comments for parsing (returns parseable code only)}
\usage{
strip_stata_comments(lines)
}
\description{
Strip Stata comments for parsing (returns parseable code only)
}
\keyword{internal}
