% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code.R
\name{strip_yaml_run_section}
\alias{strip_yaml_run_section}
\title{Drop a previously appended yaml run section from script lines}
\usage{
strip_yaml_run_section(lines)
}
\description{
Drop a previously appended yaml run section from script lines
}
\keyword{internal}
