% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_dag.R
\name{study_dag_components}
\alias{study_dag_components}
\title{Weakly connected components of analytical steps for display}
\usage{
study_dag_components(graph)
}
\description{
Weakly connected components of analytical steps for display
}
\keyword{internal}
