% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataverse_manifest.R
\name{study_dataverse_config}
\alias{study_dataverse_config}
\title{Read Dataverse config from study metadata or local replication.yml}
\usage{
study_dataverse_config(meta, ctx)
}
\description{
Read Dataverse config from study metadata or local replication.yml
}
\keyword{internal}
