% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_inheritance.R
\name{study_has_extension}
\alias{study_has_extension}
\title{Whether merged metadata includes a base study extension}
\usage{
study_has_extension(meta)
}
\description{
Whether merged metadata includes a base study extension
}
\keyword{internal}
