% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_replication.R
\name{study_package_error}
\alias{study_package_error}
\title{Error condition when a study replication package is not available}
\usage{
study_package_error(package, meta, ctx)
}
\arguments{
\item{package}{R package name.}

\item{meta}{Parsed replication metadata.}

\item{ctx}{Paper context.}
}
\description{
Error condition when a study replication package is not available
}
\keyword{internal}
