% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_links.R
\name{study_relative_path}
\alias{study_relative_path}
\title{Relative path of an absolute file within a study root}
\usage{
study_relative_path(abs_path, study_root)
}
\description{
Relative path of an absolute file within a study root
}
\keyword{internal}
