% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_inheritance.R
\name{study_repo_folder_name}
\alias{study_repo_folder_name}
\title{Repo folder name from a GitHub slug}
\usage{
study_repo_folder_name(repo_slug)
}
\description{
Repo folder name from a GitHub slug
}
\keyword{internal}
