% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_study.R
\name{study_step_type_counts}
\alias{study_step_type_counts}
\title{Count table / figure / transform steps}
\usage{
study_step_type_counts(entries)
}
\description{
Count table / figure / transform steps
}
\keyword{internal}
