% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code.R
\name{[.replication_code}
\alias{[.replication_code}
\title{Keep class after \code{head()} / subset so printing stays \code{cat()}}
\usage{
\method{[}{replication_code}(x, i, ...)
}
\description{
Keep class after \code{head()} / subset so printing stays \code{cat()}
}
\keyword{internal}
