% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substantive_checks.R
\name{substantive_check_path}
\alias{substantive_check_path}
\title{Path to a study substantive-check script}
\usage{
substantive_check_path(study_root, what)
}
\description{
Path to a study substantive-check script
}
\keyword{internal}
