% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_links.R
\name{substitute_stata_globals}
\alias{substitute_stata_globals}
\title{Substitute Stata globals in a path string}
\usage{
substitute_stata_globals(path, globals)
}
\description{
Substitute Stata globals in a path string
}
\keyword{internal}
