% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code.R
\name{suggest_data_load_expr}
\alias{suggest_data_load_expr}
\title{Suggest a load expression for a study-root-relative data path}
\usage{
suggest_data_load_expr(path)
}
\description{
Suggest a load expression for a study-root-relative data path
}
\keyword{internal}
