% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_replication.R
\name{summarize_rds_prep_output}
\alias{summarize_rds_prep_output}
\title{Summarize an RDS prep output for display}
\usage{
summarize_rds_prep_output(path, obj)
}
\arguments{
\item{path}{Local RDS path.}

\item{obj}{Object read from \code{path}.}
}
\description{
Summarize an RDS prep output for display
}
\keyword{internal}
