% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_inheritance.R
\name{tag_inherited_step}
\alias{tag_inherited_step}
\title{Mark a step as inherited from a base study}
\usage{
tag_inherited_step(step, base_repo, base_step_id)
}
\description{
Mark a step as inherited from a base study
}
\keyword{internal}
