% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_dag.R
\name{topological_step_sort}
\alias{topological_step_sort}
\title{Topological sort of a step id set respecting parent edges}
\usage{
topological_step_sort(step_ids, graph)
}
\description{
Topological sort of a step id set respecting parent edges
}
\keyword{internal}
