% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_replication.R
\name{unload_replication_package}
\alias{unload_replication_package}
\title{Unload a study replication package namespace if loaded}
\usage{
unload_replication_package(package)
}
\arguments{
\item{package}{Package name.}
}
\description{
Unload a study replication package namespace if loaded
}
\keyword{internal}
