% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_dag.R
\name{validate_given_downward_closure}
\alias{validate_given_downward_closure}
\title{Validate given is downward-closed (includes all ancestors of each given step)}
\usage{
validate_given_downward_closure(given_ids, graph)
}
\description{
Validate given is downward-closed (includes all ancestors of each given step)
}
\keyword{internal}
