% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_audit.R
\name{yaml_string_list}
\alias{yaml_string_list}
\title{Normalize a yaml list field to a character vector}
\usage{
yaml_string_list(...)
}
\description{
Normalize a yaml list field to a character vector
}
\keyword{internal}
