% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_repository.R
\name{check_paper_source_repository}
\alias{check_paper_source_repository}
\title{check_replication row for paper.source_repository}
\usage{
check_paper_source_repository(paper)
}
\arguments{
\item{paper}{Paper list.}
}
\value{
A one-row check result data frame.
}
\description{
check_replication row for paper.source_repository
}
\keyword{internal}
