% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_replication_template.R
\name{create_replication_template}
\alias{create_replication_template}
\title{Create a replication template}
\usage{
create_replication_template(doi)
}
\arguments{
\item{doi}{Character DOI of the paper.}
}
\description{
Creates a folder structure and template files for replicating a paper.
}
