% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_dag.R
\name{describe_study_dag}
\alias{describe_study_dag}
\title{Text representation of the study DAG for Shiny / CLI}
\usage{
describe_study_dag(meta)
}
\arguments{
\item{meta}{Parsed replication metadata.}
}
\value{
Character vector of component strings.
}
\description{
Text representation of the study DAG for Shiny / CLI
}
