% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_repo.R
\name{find_repo}
\alias{find_repo}
\title{Find replication repository}
\usage{
find_repo(doi)
}
\arguments{
\item{doi}{Character DOI of the paper.}
}
\description{
Returns the GitHub repository that hosts replications for a paper.
}
