% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code.R
\name{get_code}
\alias{get_code}
\title{Retrieve replication code}
\usage{
get_code(doi, what)
}
\arguments{
\item{doi}{Character DOI of the paper.}

\item{what}{Replication ID.}
}
\description{
Downloads the replication script for a specific output.
}
