% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataverse_manifest.R
\name{is_dataverse_access_prep_step}
\alias{is_dataverse_access_prep_step}
\title{Whether a prep step should use the Dataverse deposit summary display}
\usage{
is_dataverse_access_prep_step(prep, meta)
}
\description{
Whether a prep step should use the Dataverse deposit summary display
}
\keyword{internal}
