% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_replications.R
\name{list_replications}
\alias{list_replications}
\title{List available replications for a paper}
\usage{
list_replications(doi)
}
\arguments{
\item{doi}{Character. DOI of the paper.}
}
\value{
A list of replication entries from \code{replication.yml}.
}
\description{
List available replications for a paper
}
