% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_requirements.R
\name{PACKAGE_REPLICATION_API}
\alias{PACKAGE_REPLICATION_API}
\title{Required exported API for package-backed replication studies}
\usage{
PACKAGE_REPLICATION_API
}
\description{
Required exported API for package-backed replication studies
}
\keyword{internal}
