% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replication_check_print.R
\name{print.package_replication_check}
\alias{print.package_replication_check}
\alias{print.folder_replication_check}
\alias{print.replication_check}
\title{Print a replication checklist result}
\usage{
\method{print}{package_replication_check}(x, ...)

\method{print}{folder_replication_check}(x, ...)

\method{print}{replication_check}(x, ...)
}
\arguments{
\item{x}{Result from \code{\link[=check_replication]{check_replication()}}, \code{\link[=check_and_bake_study]{check_and_bake_study()}}, or
\code{\link[=register_study]{register_study()}}.}

\item{...}{Ignored.}
}
\description{
Print a replication checklist result
}
\keyword{internal}
