% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_paper.R
\name{print.package_replication_check}
\alias{print.package_replication_check}
\alias{print.folder_replication_check}
\alias{print.replication_check}
\title{Print a replication checklist result}
\usage{
\method{print}{package_replication_check}(x, ...)

\method{print}{folder_replication_check}(x, ...)

\method{print}{replication_check}(x, ...)
}
\arguments{
\item{x}{Result from \code{\link[=check_replication]{check_replication()}}, \code{\link[=add_paper]{add_paper()}}, or \code{\link[=add_folder_paper]{add_folder_paper()}}.}

\item{...}{Ignored.}
}
\description{
Print a replication checklist result
}
\keyword{internal}
