% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_paper.R
\name{print.package_replication_check}
\alias{print.package_replication_check}
\title{Print a package replication checklist result}
\usage{
\method{print}{package_replication_check}(x, ...)
}
\arguments{
\item{x}{Result from \code{\link[=check_package_replication]{check_package_replication()}} or \code{\link[=add_paper]{add_paper()}}.}

\item{...}{Ignored.}
}
\description{
Print a package replication checklist result
}
