% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_paper.R
\name{replicate_paper}
\alias{replicate_paper}
\title{Replicate an entire paper}
\usage{
replicate_paper(doi)
}
\arguments{
\item{doi}{Character DOI of the paper.}
}
\description{
Runs all replications (figures and tables) listed in the metadata.
}
