% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicateEverything-package.R
\docType{package}
\name{replicateEverything-package}
\alias{replicateEverything}
\alias{replicateEverything-package}
\title{replicateEverything: Reproduce Empirical Research Results}
\description{
The \code{replicateEverything} package provides tools for discovering and
executing computational replications of empirical research papers.
It connects to a public replication registry containing metadata,
replication scripts, and processed datasets required to reproduce
figures and tables from published studies.
}
\section{Workflow}{


A typical workflow using the package is:
\enumerate{
\item Retrieve metadata for a paper using \code{get_doi_metadata()}.
\item Search the replication registry using \code{search_papers()}.
\item Create a template for contributing a replication using
\code{create_replication_template()}.
\item Inspect available replications using \code{list_replications()}.
\item Run a single replication using \code{run_replication()}.
\item Reproduce all results from a paper using \code{replicate_paper()}.
}
}

\section{Examples}{


Retrieve metadata for a paper:

\donttest{
get_doi_metadata("10.1177/00491241211036161")
}

Search the replication registry:

\donttest{
search_papers("causal")
}

Create a template for contributing a replication:

\donttest{
create_replication_template("10.1177/00491241211036161")
}

List available replications:

\donttest{
list_replications("10.1177/00491241211036161")
}

Run a single replication:

\donttest{
run_replication("10.1177/00491241211036161", "fig_1")
}

Replicate all results from a paper:

\donttest{
replicate_paper("10.1177/00491241211036161")
}
}

\section{Registry}{


Replication metadata and materials are stored in the public registry:
\url{https://github.com/replicate-anything/registry}.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/replicate-anything/replicateEverything}
  \item Report bugs at \url{https://github.com/replicate-anything/replicateEverything/issues}
}

}
\author{
\strong{Maintainer}: Vermon Washington \email{vermon.washington@wzb.eu}

Authors:
\itemize{
  \item Macartan Humphreys \email{macartan.humphreys@wzb.eu}
}

}
\keyword{internal}
