% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_replication.R
\name{replication_package_usable}
\alias{replication_package_usable}
\title{Whether a study replication package namespace can be used}
\usage{
replication_package_usable(package)
}
\arguments{
\item{package}{Installed or dev-loaded package name.}
}
\description{
Whether a study replication package namespace can be used
}
\keyword{internal}
