% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_replication.R
\name{replication_requires_prep}
\alias{replication_requires_prep}
\title{Parent ids for a step (\verb{parents:} only)}
\usage{
replication_requires_prep(rep)
}
\arguments{
\item{rep}{Step entry.}
}
\description{
Parent ids for a step (\verb{parents:} only)
}
\keyword{internal}
