% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_papers.R
\name{search_papers}
\alias{search_papers}
\title{Search papers in the replication registry}
\usage{
search_papers(query)
}
\arguments{
\item{query}{Character search term.}
}
\description{
Searches the registry index by keyword.
}
