% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_build_info.R
\name{shiny_deploy_dir}
\alias{shiny_deploy_dir}
\title{Shiny deploy directory (app bundle root)}
\usage{
shiny_deploy_dir()
}
\value{
Normalized path.
}
\description{
Shiny deploy directory (app bundle root)
}
\keyword{internal}
