% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_inheritance.R
\name{step_run_context}
\alias{step_run_context}
\title{Paper context to use when running or checking a step}
\usage{
step_run_context(step, meta, ctx)
}
\description{
Paper context to use when running or checking a step
}
\keyword{internal}
