% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder_requirements.R
\name{study_artifact_rel_path}
\alias{study_artifact_rel_path}
\title{Expected artifact path relative to study root}
\usage{
study_artifact_rel_path(rep)
}
\description{
Expected artifact path relative to study root
}
\keyword{internal}
