% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_dag.R
\name{study_dag_display}
\alias{study_dag_display}
\title{Step display data for Shiny (components of paths of id / label / description)}
\usage{
study_dag_display(meta)
}
\arguments{
\item{meta}{Parsed replication metadata.}
}
\value{
A list of components; each component is a list of paths; each path is a
list of step display records.
}
\description{
Step display data for Shiny (components of paths of id / label / description)
}
